Rename requests_extra to librequests
